task 5 aki mungu saidia

+++ 0x0C-more_malloc_free/101-mul.c
@@ -1,11 +1,13 @@
 #include "main.h"
+#include <stdio.h>
+#include <stdlib.h>
 
 /**
- * main - multiplies two positive numbers
- * @argc: n arguments
- * @argv: args
- * Return: int
- */
+  * main - multiplies two positive numbers
+  * @argc: n arguments
+  * @argv: args
+  * Return: int
+  */
 int main(int argc, char *argv[])
 {
 unsigned long mul;
@@ -23,7 +25,7 @@ int i, j;
 		}
 
 	}
-	mul = atol(argv[1]) *  atol(argv[2]);
+	mul = atoi(argv[1]) *  atoi(argv[2]);
 	printf("%lu\n", mul);
 return (0);
 }
